tests/wpt/tests/css/css-animations/parsing/animation-delay-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing animation-delay with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-delay">
<meta name="assert" content="animation-delay supports the full grammar '<single-animation-play-state> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("animation-delay", "-5ms");
test_valid_value("animation-delay", "0s");
test_valid_value("animation-delay", "10s");
test_valid_value("animation-delay", "20s, 10s");
</script>
</body>
</html>
